+++ SAST/codeql/codeql_v2_13_1/codeql.py
@@ -7,16 +7,13 @@ DIR: Path = Path(__file__).parent.resolve()
 SAST_DIR: Path = DIR.parent.parent.resolve()
 sys.path.append(str(SAST_DIR))
 
-from codeql.core.codeql import CodeQL
+from SAST.codeql.core.codeql import CodeQL
 
-class CodeQL_v_2_9_2(CodeQL):
+class CodeQL_v_2_13_1(CodeQL):
 
     def __init__(self):
-        self.tool = "codeql_2_9_2"
+        self.tool = "codeql_2_13_1"
         self.CODEQL_SCRIPT_DIR: Path = Path(__file__).parent.resolve()
         self.CODEQL_CONFIG_FILE: Path = self.CODEQL_SCRIPT_DIR / "config.yaml"
         with open(self.CODEQL_CONFIG_FILE) as sast_config_file:
             self.CODEQL_CONFIG: Dict = yaml.load(sast_config_file, Loader=yaml.Loader)
-
-
-
