+++ SAST/codeql/core/resources/_template_build.sh
